fix(openid): Mark tests as expected failures
Long term solution: phase out OpenID 2.0 support. Current test relies on
Yahoo (dead now) and Hyves (should be long dead, if it isn't, it's
undead).

This fix silences the tests, until someone wants to find a new provider
or mock the discovery, but OpenID itself has moved on to the next shiny
version: OpenID Connect, so it is best to sunset this support.